Fix tautological-compare warning. BAD; DO NOT USE.
This warning was caused by an assertion that a `size_t` was `> 0`, which
was implemented by checking whether the variable was `< 0`, which is of
course impossible for an unsigned integer.

I changed the statement to instead assert that the variable is `!= 0`.

According to Krista, “the type was probably an integer at some point and
the test made sense”.

------------------------------------------------------------------------

…So months later Sirp makes me realize that I’ve somehow not noticed
that the assertion was actually that the `size_t` was `≥ 0`, not `> 0`.

The proper fix was thus to delete the pointless assertion entirely,
which Sirp did, in commit f79c9ce1093350f24736fe1f26a0e8abcd7a91a3.

So, DO NOT USE THIS COMMIT. IT IS BAD AND USELESS, WRITTEN UNDER A
FALSE ASSUMPTION.